+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Feat
 
 * **Integration Enviroment:** Create skaffold configuration for stagging. Closes #CSF-23
+* **Integration-Backbone:** Service to send images to Salesforce. Add endpoints to upload and associate an image to a case. Closes #CSF-60
 * **Integration-Backbone:** Create main structure. Add project  base files : in-memory store library, proxy library, add endpoint healtcheck. Closes #CSF-43
 * **Integration-Library Salesforce:** Service to long polling messages from Salesforce. Add the request to obtain the agents' messages  of a chat, the security token is added to the request to obtain token. Closes #CSF-58